backend/requirements.txt
Django
djangorestframework
django-filter
django-cors-headers
django-environ
python-dotenv
djangorestframework-simplejwt
djoser
PyJWT
psycopg2-binary
django-extensions
marshmallow
celery
redis
django-celery-beat
django-celery-results
drf-spectacular
drf-yasg
pytest
pytest-django
pytest-cov
factory-boy
faker
coverage
black
flake8
isort
pylint
django-debug-toolbar
django-ratelimit
cryptography
sentry-sdk
python-json-logger
gunicorn
whitenoise
python-dateutil
pytz
requests
numpy
pandas
